adjust windowScroll hook, update tootltip zIndexing overflow of header
